Separate average values calculations
